Added support for reading policy from the root of the mounted directory. Supports different policies for subdirectories, but will fall back to the default policy read from the root of the mounted directory. Some caveats exist in the way the Waters09 implementation handles parsing of the policy which prevents any assignments to variables to be made in the policy.
Current support for policies to containing the uid and gid of users
appended to the strings  'UID' and 'GID' respectively. Using a different
implementation of the parser or a different version of CP-ABE with a
better parser will alleviate this issue.

Fixed a bug where the size of files was being reported as 8 bytes larger
than the actual non-metadata size.

Added extra checks for padding removal to happen at the correct time.

Added ability for other users to access the FUES filesystem. If a user
is not authorized to decrypt data, then the read call for that file will
return 0 bytes. This was chosen as a simplification over returning the
ciphertext of the file.